Developed author by id functionality, altered path project.
