+++ src/main/java/com/felipemoreira/bookstore/config/SwaggerConfig.java
@@ -1,0 +1,90 @@
+package com.felipemoreira.bookstore.config;
+
+import static springfox.documentation.builders.PathSelectors.any;
+import static springfox.documentation.builders.RequestHandlerSelectors.basePackage;
+import static springfox.documentation.spi.DocumentationType.SWAGGER_2;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.List;
+import org.springframework.boot.actuate.autoconfigure.endpoint.web.CorsEndpointProperties;
+import org.springframework.boot.actuate.autoconfigure.endpoint.web.WebEndpointProperties;
+import org.springframework.boot.actuate.autoconfigure.web.server.ManagementPortType;
+import org.springframework.boot.actuate.endpoint.ExposableEndpoint;
+import org.springframework.boot.actuate.endpoint.web.EndpointLinksResolver;
+import org.springframework.boot.actuate.endpoint.web.EndpointMapping;
+import org.springframework.boot.actuate.endpoint.web.EndpointMediaTypes;
+import org.springframework.boot.actuate.endpoint.web.ExposableWebEndpoint;
+import org.springframework.boot.actuate.endpoint.web.WebEndpointsSupplier;
+import org.springframework.boot.actuate.endpoint.web.annotation.ControllerEndpointsSupplier;
+import org.springframework.boot.actuate.endpoint.web.annotation.ServletEndpointsSupplier;
+import org.springframework.boot.actuate.endpoint.web.servlet.WebMvcEndpointHandlerMapping;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.core.env.Environment;
+import org.springframework.util.StringUtils;
+import springfox.documentation.builders.ApiInfoBuilder;
+import springfox.documentation.service.ApiInfo;
+import springfox.documentation.service.Contact;
+import springfox.documentation.spring.web.plugins.Docket;
+
+@Configuration
+public class SwaggerConfig {
+
+    private static final String BASE_PACKAGE = "com.felipemoreira.bookstore";
+    private static final String API_TITLE = "Bookstore Manager Project";
+    private static final String API_DESCRIPTION = "Bookstore Manager API Professional";
+    private static final String CONTACT_NAME = "Felipe Moreira";
+    private static final String CONTACT_GITHUB = "https://github.com/FM-007";
+    private static final String CONTACT_EMAIL = "felipe@test.com";
+    private static final String VERSION_API = "1.0.0";
+
+    @Bean
+    public Docket api() {
+        return new Docket(SWAGGER_2)
+            .select()
+            .apis(basePackage(BASE_PACKAGE))
+            .paths(any())
+            .build().apiInfo(buildApiInfo());
+    }
+
+    private ApiInfo buildApiInfo() {
+        return new ApiInfoBuilder()
+            .title(API_TITLE)
+            .description(API_DESCRIPTION)
+            .version(VERSION_API)
+            .contact(new Contact(CONTACT_NAME, CONTACT_GITHUB, CONTACT_EMAIL))
+            .build();
+    }
+
+    @Bean
+    public WebMvcEndpointHandlerMapping webEndpointServletHandlerMapping(
+        WebEndpointsSupplier webEndpointsSupplier,
+        ServletEndpointsSupplier servletEndpointsSupplier,
+        ControllerEndpointsSupplier controllerEndpointsSupplier,
+        EndpointMediaTypes endpointMediaTypes, CorsEndpointProperties corsProperties,
+        WebEndpointProperties webEndpointProperties, Environment environment) {
+
+        List<ExposableEndpoint<?>> allEndpoints = new ArrayList<>();
+        Collection<ExposableWebEndpoint> webEndpoints = webEndpointsSupplier.getEndpoints();
+        allEndpoints.addAll(webEndpoints);
+        allEndpoints.addAll(servletEndpointsSupplier.getEndpoints());
+        allEndpoints.addAll(controllerEndpointsSupplier.getEndpoints());
+
+        String basePath = webEndpointProperties.getBasePath();
+        EndpointMapping endpointMapping = new EndpointMapping(basePath);
+        boolean shouldRegisterLinksMapping = this.shouldRegisterLinksMapping(webEndpointProperties,
+            environment, basePath);
+
+        return new WebMvcEndpointHandlerMapping(endpointMapping, webEndpoints, endpointMediaTypes,
+            corsProperties.toCorsConfiguration(), new EndpointLinksResolver(allEndpoints, basePath),
+            shouldRegisterLinksMapping, null);
+    }
+
+
+    private boolean shouldRegisterLinksMapping(WebEndpointProperties webEndpointProperties,
+        Environment environment, String basePath) {
+        return webEndpointProperties.getDiscovery().isEnabled() && (StringUtils.hasText(basePath)
+            || ManagementPortType.get(environment).equals(ManagementPortType.DIFFERENT));
+    }
+}
