+++ src/main/java/com/felipemoreira/bookstore/domain/exception/ApiError.java
