+++ src/main/java/com/felipemoreira/bookstore/domain/exception/AuthorAlreadyExistsException.java
