+++ src/main/java/com/felipemoreira/bookstore/domain/exception/AuthorNotFoundException.java
@@ -1,0 +1,10 @@
+package com.felipemoreira.bookstore.domain.exception;
+
+import javax.persistence.EntityNotFoundException;
+
+public class AuthorNotFoundException extends EntityNotFoundException {
+
+    public AuthorNotFoundException(Long id) {
+        super(String.format("Author with id %s not exists!", id));
+    }
+}
