+++ src/main/java/com/felipemoreira/bookstore/domain/exception/BookstoreExceptionHandler.java
