+++ src/main/java/com/felipemoreira/bookstore/entities/Auditable.java
