+++ src/main/resources/application-dev.properties
@@ -1,0 +1,24 @@
+####### Data Base
+spring.datasource.url=jdbc:h2:mem:bookstore_dev;DB_CLOSE_DELAY=-1;DB_CLOSE_ON_EXIT=FALSE
+spring.datasource.username=sa
+spring.datasource.password=
+spring.datasource.driver-class-name=org.h2.Driver
+
+####### Spring JPA
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
+spring.jpa.show-sql=true
+spring.jpa.properties.hibernate.format_sql=true
+
+#spring.jpa.open-in-view=false
+#spring.jpa.properties.hibernate.enable_lazy_load_no_trans=false
+#spring.jpa.properties.hibernate.default_schema=BOOKSTORE
+spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.properties.hibernate.jdbc.time_zone=America/Sao_Paulo
+#spring.jpa.properties.hibernate.jdbc.batch_size=30
+#spring.jpa.properties.hibernate.order_inserts=true
+#spring.jpa.properties.hibernate.generate_statistics=false
+
+####### H2
+spring.h2.console.enabled=true
+
+logging.level.br.com.felipemoreira= DEBUG
