+++ src/test/java/com/felipemoreira/bookstore/utils/constants/AuthorConstants.java
@@ -1,0 +1,6 @@
+package com.felipemoreira.bookstore.utils.constants;
+
+public class AuthorConstants {
+
+    public static final String AUTHOR_URL = "/api/v1/authors";
+}
